+++ app/views/favorites/_show.html.erb
@@ -1,13 +1,15 @@
-<turbo-frame id="favorite-btn<%= post.id %>">
-	<% if user_signed_in? %>
+<% if user_signed_in? %>
+	<turbo-frame id="favorite-btn<%= post.id %><%= local_assigns[:flag] %>">
 		<% if post.favorited_by?(current_user)%>
-			<%= button_to post_favorite_path(post, id: post.favorites), method: :delete do %>
+			<%= button_to post_favorite_path(post, id: post.favorites, flag: local_assigns[:flag]), method: :delete do %>
 				<i class="fa-solid fa-heart"></i>
+				<%= post.favorites.count %>
 			<% end %>
 		<% else %>
-			<%= button_to post_favorites_path(post), method: :post  do %>
+			<%= button_to post_favorites_path(post, flag: local_assigns[:flag]), method: :post  do %>
 				<i class="fa-regular fa-heart"></i>
+				<%= post.favorites.count %>
 			<% end %>
 		<% end %>
-	<% end %>
-</turbo-frame>
+	</turbo-frame>
+<% end %>
