+++ app/views/posts/_favlists.html.erb
@@ -1,0 +1,3 @@
+<div class="contents" id="fav-lists">
+	<div class="contents-item"><%= render partial: "show", collection: @favorite_posts, as: "post" %></div>
+</div>
