+++ app/views/posts/_show.html.erb
@@ -3,5 +3,5 @@
 	[<%= post.user.id %>]<%= post.user.name %>."<%= post.text %>",<%= post.comments.count%>,♡<%= post.favorites.count %>
 	<%#<%= render "favorites/show", {post: post} %>
 	<% end %>
-	<%= render "favorites/show", {post: post} %>
+	<%= render "favorites/show", {post: post, flag: flag} %>
 </li>
